fix: resolve narrowing conversion warnings (bugprone-narrowing-conversions)
Applied safety fixes to prevent implicit type conversions that could cause
data loss or unexpected behavior:

- manager.cc: Use double for atof() result (line 2969), size_t for strlen
  results (lines 3088-3089) instead of int
- zone.cc: Add explicit static_cast<short> for Page.size assignment (1274),
  static_cast<int> for strlen (1496)
- user_edit_zone.cc: Add explicit static_cast<short> for FormField::Set()
  call with int parameter (739)

These changes follow clang-tidy bugprone-narrowing-conversions recommendations
and improve type safety across the codebase.

Verification: Build successful, all 40 tests passing.
Updated changelog.md to document these safety improvements.

+++ main/data/manager.cc
@@ -2966,9 +2966,9 @@ int DeliveryToInt(const char* cost)
 {
     FnTrace("DeliveryToInt()");
     int retval = 0;
-    float interm = atof(cost);
+    double interm = atof(cost);
 
-    retval = (int)(interm * 100.0);
+    retval = static_cast<int>(interm * 100.0);
 
     return retval;
 }
@@ -3080,8 +3080,8 @@ int CompareCardNumbers(const char* card1, const char* card2)
 {
     FnTrace("CompreCardNumbers()");
     int retval = 0;
-    int len1 = 0;
-    int len2 = 0;
+    size_t len1 = 0;
+    size_t len2 = 0;
 
     if (card1[0] == 'x' || card2[0] == 'x')
     {
